collector/requirements.txt
protobuf>=3.0.0
grpcio>=1.21.1
mysql-connector-python>=8.0